Add searchbar style
