--- src/client/components/top_bar/style.css
@@ -1,4 +1,0 @@
-.searchBarId {
-  height:1000px;
-  font-size:14pt;
-}
