+++ src/client/pages/ProductListPage.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import TopBar from '../components/top_bar/TopBar.js';
-import ProductList from '../components/product_list/ProductList.js';
+import ProductGridList from '../components/product_list/ProductGridList.js';
 
 class ProductListPage extends Component {
   render() {
     return (
       <div>
         <TopBar/>
-        <ProductList/>
+        <ProductGridList/>
       </div>
     );
   }
